services/blinkpdf/src/requirements.txt
flask==2.3.3
ecdsa==0.19.0
PyPDF2==3.0.1
python-dotenv==1.0.1